+++ lib/checkbox_grouped.dart
@@ -5,3 +5,4 @@ export 'src/custom_grouped_checkbox.dart';
 export 'src/simple_grouped_chips.dart';
 export 'src/simple_grouped_switch.dart';
 export 'src/show_dialog_grouped_checkbox.dart';
+export 'src/list_grouped_checkbox.dart';